Genomic_interval_collection: more general type for intersection
in particular, refrain from computing a dictionary immediately, and
let's see how far we can go with a simple assoc list.

in GREAT mode, constructing the dictionary is not necessary at all if
we don't write optional outputs. Then this change brings a slight
speed up (from 13s to 12s).

However, I had to change a lot of unprofiled code for contact and
hybrid mode so it might have get worse there. Let's hope I won't need
to backtrack on this!